requirements.txt
cycler==0.10.0
kiwisolver==1.0.1
matplotlib==3.0.3
numpy==1.16.2
opencv-contrib-python==3.4.2.16
pyparsing==2.3.1
python-dateutil==2.8.0
python-dotenv==0.10.1
six==1.12.0
